Add CoFHCore/Lib as a dependency to the build system
The mod will require CoFHCore/Lib as a main dependency, so I added it to the
build system. The method used is not the best implementation, which is adding
it to the main source set during development, and stripping it out from the
final obfuscated jar, but after much experimentation, this was the easiest
method found with the least amount of problems occuring. During build, the CoFH
libraries are cloned from github using the CoFHCore/Lib_URL and _HASH
properties, as CoFHCore gets upgraded, this hash will change to point to a new
commit. The resulting files are cloned within src/cofh/[core, lib], and their
source added to the main project.