Tweak Configuration gui to hide subcategories from the main view
The configuration gui will now not show subcategories from the main config
screen, and now is able to be drilled down as expected through the various
subcategories. Also renamed a few files to bring them more in line with the way
other CoFH mods do the naming of their Gui configurations.