Fix broken checkout that occurs when updating CoFHCore/Lib
The checkout for the CoFHCore git repository failed because I make
modifications to the git repository in order to strip the mcmod.info file from
the classpath. I do this so that the mcmod.info file does not overwrite my own,
and since it works fine without it. This is done by having the git repository
hard reset before the checkout occurs. This is fine as the library should never
be changed within the project itself.