Tweak Config system to enable correct handling of nested categories
The configuration system will now correctly show nested categories within the
configuration gui, and will also localize the category names depending on the
language file present. Also a change to the config file to use a single
Astronomia.cfg file within the root of the config directory, rather than one in
it's own directory, this may change in the future if necessary for additional
functionality.