Add ModItemRegistry class, a container class for mod items/blocks
Added a new class, the ModItemRegistry, which includes a set of helper
functions and is intended to be used to contain the various items and blocks
that a mod has. There are a couple helper functions added to facilitate easier
adding and initializing the mods and items. This also has the added benefit of
automatically adding it to the configuration file with added localizations for
the configuration gui.