Add the first item to the game
This first item will be the base for the start of the research system. It will
be used for looking up into the night sky so you can make drawings of the stars
and other celestial objects. Right now it doesn't do anything, just having
placeholder functionality, and having no texture, but it will take some time to
work out how to get it all working.