Sync to latest CoFHCore/Lib and update to forge 1205
The latest CoFHCore/Lib requires forge 1205, so updating to that version, and
also syncing to them as well. Also the development will in the future be going
at a slightly different methodology, there will now be a develop branch which
will be getting main development, and when a release is near, will merge into
the master branch as well as tagging the release. Development also will be
hopefully be getting committed in discrete chunks, and more as development
progresses.